Adding Security #3
SOLUTION NOT OPTIMAL FOR UPDATE AND DELETE

As in previous SOLUTION NOT OPTIMAL commit for CREATE, here again, the "auth-interceptor" class is deactivated as it does not send the HttpHeader with the auth token to methods at request (although the right token info is in the Session Storage, the value sent is always "Bearer undefined").

Added "en dur" to methods CREATE and DELETE the token always valid of one particular user, so:
- Now if we create/update/delete one pixelart, the operations are controled (by the Back-end) as normally for that particular user.

The control functions established in the Back-end can be demonstrated. :)